Add cache to Athena start query
